+++ models/organization_test.go
@@ -6,7 +6,7 @@ import (
 )
 
 func TestOrganizationModel(t *testing.T) {
-    model := Organization{        Id: "5e5ea5c16897e",        CreatedAt: "2020-10-15T06:38:00.000+00:00",        UpdatedAt: "2020-10-15T06:38:00.000+00:00",        Name: "VIP",        Total: 7,        Prefs: Preferences{    },        BillingBudget: 50,        BudgetAlerts: []int{1},        BillingPlan: "tier-1",        BillingPlanId: "tier-1",        BillingPlanDetails: BillingPlan{        Id: "tier-0",        Name: "Hobby",        Desc: "Hobby plan",        Order: 0,        Price: 25,        Trial: 14,        Bandwidth: 25,        Storage: 25,        ImageTransformations: 100,        ScreenshotsGenerated: 50,        Members: 25,        Webhooks: 25,        WafRules: 2,        Projects: 2,        Platforms: 3,        Users: 25,        Teams: 25,        Databases: 25,        DatabasesReads: 500000,        DatabasesWrites: 250000,        DatabasesBatchSize: 100,        Buckets: 25,        FileSize: 25,        Functions: 25,        Sites: 1,        Executions: 25,        ExecutionsRetentionCount: 10000,        GBHours: 100,        Realtime: 25,        RealtimeMessages: 100000,        Messages: 1000,        Topics: 1,        AuthPhone: 10,        Domains: 5,        ActivityLogs: 7,        UsageLogs: 30,        ProjectInactivityDays: 7,        AlertLimit: 80,        Usage: UsageBillingPlan{        Bandwidth: AdditionalResource{        Name: "string",        Unit: "GB",        Currency: "USD",        Price: 5,        Value: 25,        InvoiceDesc: "string",    },        Executions: AdditionalResource{        Name: "string",        Unit: "GB",        Currency: "USD",        Price: 5,        Value: 25,        InvoiceDesc: "string",    },        Member: AdditionalResource{        Name: "string",        Unit: "GB",        Currency: "USD",        Price: 5,        Value: 25,        InvoiceDesc: "string",    },        Realtime: AdditionalResource{        Name: "string",        Unit: "GB",        Currency: "USD",        Price: 5,        Value: 25,        InvoiceDesc: "string",    },        RealtimeMessages: AdditionalResource{        Name: "string",        Unit: "GB",        Currency: "USD",        Price: 5,        Value: 25,        InvoiceDesc: "string",    },        RealtimeBandwidth: AdditionalResource{        Name: "string",        Unit: "GB",        Currency: "USD",        Price: 5,        Value: 25,        InvoiceDesc: "string",    },        Storage: AdditionalResource{        Name: "string",        Unit: "GB",        Currency: "USD",        Price: 5,        Value: 25,        InvoiceDesc: "string",    },        Users: AdditionalResource{        Name: "string",        Unit: "GB",        Currency: "USD",        Price: 5,        Value: 25,        InvoiceDesc: "string",    },        GBHours: AdditionalResource{        Name: "string",        Unit: "GB",        Currency: "USD",        Price: 5,        Value: 25,        InvoiceDesc: "string",    },        ImageTransformations: AdditionalResource{        Name: "string",        Unit: "GB",        Currency: "USD",        Price: 5,        Value: 25,        InvoiceDesc: "string",    },        Credits: AdditionalResource{        Name: "string",        Unit: "GB",        Currency: "USD",        Price: 5,        Value: 25,        InvoiceDesc: "string",    },    },        Addons: BillingPlanAddon{        Seats: BillingPlanAddonDetails{        Supported: true,        PlanIncluded: 1,        Limit: 5,        Type: "numeric",        Currency: "USD",        Price: 5,        Value: 25,        InvoiceDesc: "string",    },        Projects: BillingPlanAddonDetails{        Supported: true,        PlanIncluded: 1,        Limit: 5,        Type: "numeric",        Currency: "USD",        Price: 5,        Value: 25,        InvoiceDesc: "string",    },    },        BudgetCapEnabled: true,        CustomSmtp: true,        EmailBranding: true,        RequiresPaymentMethod: true,        RequiresBillingAddress: true,        IsAvailable: true,        SelfService: true,        PremiumSupport: true,        Budgeting: true,        SupportsMockNumbers: true,        SupportsOrganizationRoles: true,        SupportsCredits: true,        SupportsDisposableEmailValidation: true,        SupportsCanonicalEmailValidation: true,        SupportsFreeEmailValidation: true,        SupportsCorporateEmailValidation: true,        SupportsProjectSpecificRoles: true,        BackupsEnabled: true,        UsagePerProject: true,        SupportedAddons: BillingPlanSupportedAddons{        Baa: true,        PremiumGeoDB: true,        PremiumGeoDBOrg: true,    },        BackupPolicies: 1,        DeploymentSize: 30,        BuildSize: 2000,        DatabasesAllowEncrypt: true,        Group: "pro",    },        BillingEmail: "billing@org.example",        BillingStartDate: "2020-10-15T06:38:00.000+00:00",        BillingCurrentInvoiceDate: "2020-10-15T06:38:00.000+00:00",        BillingNextInvoiceDate: "2020-10-15T06:38:00.000+00:00",        BillingTrialStartDate: "2020-10-15T06:38:00.000+00:00",        BillingTrialDays: 14,        BillingAggregationId: "adbc3de4rddfsd",        BillingInvoiceId: "adbc3de4rddfsd",        PaymentMethodId: "adbc3de4rddfsd",        BillingAddressId: "adbc3de4rddfsd",        BackupPaymentMethodId: "adbc3de4rddfsd",        Status: "active",        Remarks: "Pending initial payment",        AgreementBAA: "string",        ProgramManagerName: "string",        ProgramManagerCalendar: "string",        ProgramDiscordChannelName: "string",        ProgramDiscordChannelUrl: "string",        BillingPlanDowngrade: "tier-1",        BillingTaxId: "string",        MarkedForDeletion: true,        Platform: "imagine",        Projects: []string{"test"},    }
+    model := Organization{        Id: "5e5ea5c16897e",        CreatedAt: "2020-10-15T06:38:00.000+00:00",        UpdatedAt: "2020-10-15T06:38:00.000+00:00",        Name: "VIP",        Total: 7,        Prefs: Preferences{    },        BudgetAlerts: []int{1},        BillingPlan: "tier-1",        BillingPlanId: "tier-1",        BillingPlanDetails: BillingPlan{        Id: "tier-0",        Name: "Hobby",        Desc: "Hobby plan",        Order: 0,        Price: 25,        Trial: 14,        Bandwidth: 25,        Storage: 25,        ImageTransformations: 100,        ScreenshotsGenerated: 50,        Webhooks: 25,        WafRules: 2,        Projects: 2,        Platforms: 3,        Users: 25,        Teams: 25,        Databases: 25,        DatabasesReads: 500000,        DatabasesWrites: 250000,        DatabasesBatchSize: 100,        Buckets: 25,        FileSize: 25,        Functions: 25,        Sites: 1,        Executions: 25,        ExecutionsRetentionCount: 10000,        GBHours: 100,        Realtime: 25,        RealtimeMessages: 100000,        Messages: 1000,        Topics: 1,        AuthPhone: 10,        Domains: 5,        UsageLogs: 30,        ProjectInactivityDays: 7,        AlertLimit: 80,        Usage: UsageBillingPlan{        Bandwidth: AdditionalResource{        Name: "string",        Unit: "GB",        Currency: "USD",        Price: 5,        Value: 25,        InvoiceDesc: "string",    },        Executions: AdditionalResource{        Name: "string",        Unit: "GB",        Currency: "USD",        Price: 5,        Value: 25,        InvoiceDesc: "string",    },        Realtime: AdditionalResource{        Name: "string",        Unit: "GB",        Currency: "USD",        Price: 5,        Value: 25,        InvoiceDesc: "string",    },        RealtimeMessages: AdditionalResource{        Name: "string",        Unit: "GB",        Currency: "USD",        Price: 5,        Value: 25,        InvoiceDesc: "string",    },        Storage: AdditionalResource{        Name: "string",        Unit: "GB",        Currency: "USD",        Price: 5,        Value: 25,        InvoiceDesc: "string",    },        Users: AdditionalResource{        Name: "string",        Unit: "GB",        Currency: "USD",        Price: 5,        Value: 25,        InvoiceDesc: "string",    },        GBHours: AdditionalResource{        Name: "string",        Unit: "GB",        Currency: "USD",        Price: 5,        Value: 25,        InvoiceDesc: "string",    },        ImageTransformations: AdditionalResource{        Name: "string",        Unit: "GB",        Currency: "USD",        Price: 5,        Value: 25,        InvoiceDesc: "string",    },    },        Addons: BillingPlanAddon{    },        BudgetCapEnabled: true,        CustomSmtp: true,        EmailBranding: true,        RequiresPaymentMethod: true,        RequiresBillingAddress: true,        IsAvailable: true,        SelfService: true,        PremiumSupport: true,        Budgeting: true,        SupportsMockNumbers: true,        SupportsOrganizationRoles: true,        SupportsCredits: true,        SupportsDisposableEmailValidation: true,        SupportsCanonicalEmailValidation: true,        SupportsFreeEmailValidation: true,        SupportsCorporateEmailValidation: true,        SupportsProjectSpecificRoles: true,        UsagePerProject: true,        SupportedAddons: BillingPlanSupportedAddons{        Baa: true,        PremiumGeoDB: true,        PremiumGeoDBOrg: true,    },        DeploymentSize: 30,        BuildSize: 2000,        DatabasesAllowEncrypt: true,        Group: "pro",    },        BillingEmail: "billing@org.example",        BillingStartDate: "2020-10-15T06:38:00.000+00:00",        BillingCurrentInvoiceDate: "2020-10-15T06:38:00.000+00:00",        BillingNextInvoiceDate: "2020-10-15T06:38:00.000+00:00",        BillingTrialDays: 14,        BillingAggregationId: "adbc3de4rddfsd",        BillingInvoiceId: "adbc3de4rddfsd",        PaymentMethodId: "adbc3de4rddfsd",        Status: "active",        MarkedForDeletion: true,        Platform: "imagine",        Projects: []string{"test"},    }
 
     data, err := json.Marshal(model)
     if err != nil {
@@ -33,9 +33,6 @@ func TestOrganizationModel(t *testing.T) {
     if result.Total != model.Total {
         t.Errorf("Expected Total %v, got %v", model.Total, result.Total)
     }
-    if result.BillingBudget != model.BillingBudget {
-        t.Errorf("Expected BillingBudget %v, got %v", model.BillingBudget, result.BillingBudget)
-    }
     if result.BillingPlan != model.BillingPlan {
         t.Errorf("Expected BillingPlan %v, got %v", model.BillingPlan, result.BillingPlan)
     }
@@ -54,9 +51,6 @@ func TestOrganizationModel(t *testing.T) {
     if result.BillingNextInvoiceDate != model.BillingNextInvoiceDate {
         t.Errorf("Expected BillingNextInvoiceDate %v, got %v", model.BillingNextInvoiceDate, result.BillingNextInvoiceDate)
     }
-    if result.BillingTrialStartDate != model.BillingTrialStartDate {
-        t.Errorf("Expected BillingTrialStartDate %v, got %v", model.BillingTrialStartDate, result.BillingTrialStartDate)
-    }
     if result.BillingTrialDays != model.BillingTrialDays {
         t.Errorf("Expected BillingTrialDays %v, got %v", model.BillingTrialDays, result.BillingTrialDays)
     }
@@ -69,38 +63,8 @@ func TestOrganizationModel(t *testing.T) {
     if result.PaymentMethodId != model.PaymentMethodId {
         t.Errorf("Expected PaymentMethodId %v, got %v", model.PaymentMethodId, result.PaymentMethodId)
     }
-    if result.BillingAddressId != model.BillingAddressId {
-        t.Errorf("Expected BillingAddressId %v, got %v", model.BillingAddressId, result.BillingAddressId)
-    }
-    if result.BackupPaymentMethodId != model.BackupPaymentMethodId {
-        t.Errorf("Expected BackupPaymentMethodId %v, got %v", model.BackupPaymentMethodId, result.BackupPaymentMethodId)
-    }
     if result.Status != model.Status {
         t.Errorf("Expected Status %v, got %v", model.Status, result.Status)
-    }
-    if result.Remarks != model.Remarks {
-        t.Errorf("Expected Remarks %v, got %v", model.Remarks, result.Remarks)
-    }
-    if result.AgreementBAA != model.AgreementBAA {
-        t.Errorf("Expected AgreementBAA %v, got %v", model.AgreementBAA, result.AgreementBAA)
-    }
-    if result.ProgramManagerName != model.ProgramManagerName {
-        t.Errorf("Expected ProgramManagerName %v, got %v", model.ProgramManagerName, result.ProgramManagerName)
-    }
-    if result.ProgramManagerCalendar != model.ProgramManagerCalendar {
-        t.Errorf("Expected ProgramManagerCalendar %v, got %v", model.ProgramManagerCalendar, result.ProgramManagerCalendar)
-    }
-    if result.ProgramDiscordChannelName != model.ProgramDiscordChannelName {
-        t.Errorf("Expected ProgramDiscordChannelName %v, got %v", model.ProgramDiscordChannelName, result.ProgramDiscordChannelName)
-    }
-    if result.ProgramDiscordChannelUrl != model.ProgramDiscordChannelUrl {
-        t.Errorf("Expected ProgramDiscordChannelUrl %v, got %v", model.ProgramDiscordChannelUrl, result.ProgramDiscordChannelUrl)
-    }
-    if result.BillingPlanDowngrade != model.BillingPlanDowngrade {
-        t.Errorf("Expected BillingPlanDowngrade %v, got %v", model.BillingPlanDowngrade, result.BillingPlanDowngrade)
-    }
-    if result.BillingTaxId != model.BillingTaxId {
-        t.Errorf("Expected BillingTaxId %v, got %v", model.BillingTaxId, result.BillingTaxId)
     }
     if result.MarkedForDeletion != model.MarkedForDeletion {
         t.Errorf("Expected MarkedForDeletion %v, got %v", model.MarkedForDeletion, result.MarkedForDeletion)